+++ infer/platform_linux.c
@@ -11,8 +11,7 @@ void sleep_in_ms(uint32_t ms) {
     usleep(ms * 1000);
 }
 
-// NOTE 返回的时间戳是32位的，存在2038问题！
-uint32_t get_timestamp_in_ms() {
+uint64_t get_timestamp_in_ms() {
     struct timespec time;
     clock_gettime(CLOCK_REALTIME, &time);
     return time.tv_sec * 1000 + time.tv_nsec / 1000000;
